refactor(settings): lift the theme editor out of the screen
Why: ~750 of Settings.tsx's 2,063 lines were theme editing, which is a
feature of its own and not part of the screen's composition. Pure move —
no behaviour change.
