+++ src/screens/Settings.tsx
@@ -13,14 +13,11 @@ import {
 import {
   BUILTIN_THEMES,
   DENSITY_STEP_PX,
-  THEME_COLOR_FIELDS,
   ZOOM_MAX,
   ZOOM_MIN,
-  applyTheme,
   isValidDiffToolName,
   useSettingsStore,
   type SettingsImportReport,
-  type ThemeColors,
   type ThemeDef,
   type ThemeFollowMode,
   type UpdateChannel,
@@ -41,6 +38,7 @@ import {
   SettingsCard,
   SettingsRow,
 } from "@/features/settings/layout/SettingsCard";
+import { ThemeEditorDialog } from "@/features/settings/theme/ThemeEditorDialog";
 import { ForgeSettings } from "@/features/forge/ForgeSettings";
 import {
   cliShimStatus,
@@ -1524,478 +1522,3 @@ function AppearanceSection({ active }: { active: ThemeDef }) {
     </SettingsCard>
   );
 }
-
-// ═════════════════════════════════════════════════════════════════════════════
-// THEME EDITOR DIALOG — color pickers live here, behind Add/Edit buttons.
-// ═════════════════════════════════════════════════════════════════════════════
-
-function ThemeEditorDialog({
-  mode,
-  sourceTheme,
-  onClose,
-}: {
-  mode: "new" | "edit";
-  sourceTheme: ThemeDef;
-  onClose: () => void;
-}) {
-  const [name, setName] = React.useState(
-    mode === "new" ? `${sourceTheme.name} (custom)` : sourceTheme.name,
-  );
-  const [themeMode, setThemeMode] = React.useState<"dark" | "light">(sourceTheme.mode);
-  const [colors, setColors] = React.useState<ThemeColors>({ ...sourceTheme.colors });
-
-  // Capture the theme that was active on open, so Cancel can restore it.
-  const originalActiveRef = React.useRef<ThemeDef | null>(null);
-  if (originalActiveRef.current === null) {
-    originalActiveRef.current = useSettingsStore.getState().getActiveTheme();
-  }
-
-  // Live preview: apply draft to CSS vars whenever it changes.
-  React.useEffect(() => {
-    applyTheme({
-      id: "__draft__",
-      name,
-      mode: themeMode,
-      colors,
-    });
-  }, [name, themeMode, colors]);
-
-  // Close on Escape.
-  React.useEffect(() => {
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === "Escape") handleCancel();
-    };
-    window.addEventListener("keydown", onKey);
-    return () => window.removeEventListener("keydown", onKey);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const handleCancel = () => {
-    // Restore the theme that was active before the dialog opened.
-    const orig = originalActiveRef.current;
-    if (orig) applyTheme(orig);
-    onClose();
-  };
-
-  const handleSave = () => {
-    const trimmed = name.trim();
-    if (!trimmed) {
-      pgFlash("Theme name can't be empty");
-      return;
-    }
-    const store = useSettingsStore.getState();
-    if (mode === "new") {
-      // Create a new custom theme, make it active.
-      // First switch to source so saveAsNewTheme copies right colors — but we
-      // have our own draft, so inject directly.
-      const created = store.saveAsNewTheme(trimmed);
-      // Overwrite its colors/mode with the draft.
-      useSettingsStore.setState((st) => ({
-        customThemes: st.customThemes.map((t) =>
-          t.id === created.id
-            ? { ...t, name: trimmed, mode: themeMode, colors: { ...colors } }
-            : t,
-        ),
-      }));
-      // Through the store, not just `setState`: the draft's dark/light toggle
-      // may have flipped the mode the duplicate inherited, and in "follow the
-      // system" mode that decides WHICH half of the pairing this theme is.
-      useSettingsStore.getState().setActiveThemeId(created.id);
-      // Re-apply so the saved version is what's showing.
-      applyTheme({
-        ...created,
-        name: trimmed,
-        mode: themeMode,
-        colors,
-      });
-      pgFlash(`Saved "${trimmed}"`);
-    } else {
-      // Edit existing custom theme.
-      useSettingsStore.setState((st) => ({
-        customThemes: st.customThemes.map((t) =>
-          t.id === sourceTheme.id
-            ? { ...t, name: trimmed, mode: themeMode, colors: { ...colors } }
-            : t,
-        ),
-      }));
-      useSettingsStore.getState().setActiveThemeId(sourceTheme.id);
-      applyTheme({
-        ...sourceTheme,
-        name: trimmed,
-        mode: themeMode,
-        colors,
-      });
-      pgFlash(`Saved "${trimmed}"`);
-    }
-    onClose();
-  };
-
-  const handleResetColors = () => {
-    setColors({ ...sourceTheme.colors });
-    setThemeMode(sourceTheme.mode);
-  };
-
-  const patch = (p: Partial<ThemeColors>) =>
-    setColors((c) => ({ ...c, ...p }));
-
-  return (
-    <div
-      onClick={handleCancel}
-      style={{
-        position: "fixed",
-        inset: 0,
-        background: "rgba(0,0,0,0.55)",
-        backdropFilter: "blur(2px)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        zIndex: 100,
-        padding: 24,
-      }}
-    >
-      <div
-        onClick={(e) => e.stopPropagation()}
-        role="dialog"
-        aria-modal="true"
-        aria-label={mode === "new" ? "New custom theme" : "Edit custom theme"}
-        style={{
-          width: "min(860px, 100%)",
-          maxHeight: "100%",
-          display: "flex",
-          flexDirection: "column",
-          background: "var(--bg-1)",
-          border: "1px solid var(--border-1)",
-          borderRadius: "var(--r-5)",
-          boxShadow: "var(--shadow-3)",
-          overflow: "hidden",
-        }}
-      >
-        <header
-          style={{
-            display: "flex",
-            alignItems: "center",
-            gap: 10,
-            padding: "12px 16px",
-            borderBottom: "1px solid var(--border-0)",
-            background: "var(--bg-2)",
-          }}
-        >
-          <PGIcon name="edit" size={14} style={{ color: "var(--accent)" }} />
-          <div
-            style={{
-              fontFamily: "var(--font-mono)",
-              fontSize: "var(--fs-11)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              color: "var(--fg-1)",
-              fontWeight: 600,
-            }}
-          >
-            {mode === "new" ? "New custom theme" : "Edit custom theme"}
-          </div>
-          <div style={{ flex: 1 }} />
-          <PGButton size="sm" variant="ghost" onClick={handleResetColors}>
-            Revert changes
-          </PGButton>
-        </header>
-
-        <div style={{ padding: "14px 16px", borderBottom: "1px solid var(--border-0)" }}>
-          <div
-            style={{
-              display: "flex",
-              gap: 12,
-              alignItems: "center",
-              flexWrap: "wrap",
-            }}
-          >
-            <label
-              style={{
-                display: "flex",
-                alignItems: "center",
-                gap: 8,
-                flex: 1,
-                minWidth: 240,
-              }}
-            >
-              <span
-                style={{
-                  fontSize: "var(--fs-12)",
-                  color: "var(--fg-2)",
-                  width: 48,
-                  flexShrink: 0,
-                }}
-              >
-                Name
-              </span>
-              <PGInput
-                value={name}
-                onChange={setName}
-                placeholder="My cool theme"
-                style={{ flex: 1 }}
-              />
-            </label>
-            <div style={{ display: "flex", alignItems: "center", gap: 8 }}>
-              <span style={{ fontSize: "var(--fs-12)", color: "var(--fg-2)" }}>
-                Mode
-              </span>
-              <PGButtonGroup
-                size="sm"
-                value={themeMode}
-                onChange={(v) => setThemeMode(v as "dark" | "light")}
-                options={[
-                  { value: "dark", label: "Dark" },
-                  { value: "light", label: "Light" },
-                ]}
-              />
-            </div>
-          </div>
-          <div
-            style={{
-              marginTop: 6,
-              fontSize: "var(--fs-11)",
-              color: "var(--fg-3)",
-            }}
-          >
-            Changes preview live. Cancel to discard, Save to keep.
-          </div>
-        </div>
-
-        <div style={{ flex: 1, minHeight: 0, overflow: "auto" }}>
-          <ColorEditor colors={colors} onPatch={patch} />
-        </div>
-
-        <footer
-          style={{
-            padding: "10px 16px",
-            borderTop: "1px solid var(--border-0)",
-            display: "flex",
-            gap: 6,
-            alignItems: "center",
-            background: "var(--bg-2)",
-          }}
-        >
-          <PGButton
-            size="sm"
-            variant="default"
-            icon="download"
-            onClick={() => {
-              // Export the current draft without saving.
-              const payload = JSON.stringify(
-                {
-                  $schema: "https://platypusgit.dev/theme.schema.json",
-                  version: 1,
-                  name,
-                  mode: themeMode,
-                  colors,
-                },
-                null,
-                2,
-              );
-              const slug = (name || "theme")
-                .toLowerCase()
-                .replace(/[^a-z0-9]+/g, "-")
-                .replace(/(^-|-$)/g, "");
-              const blob = new Blob([payload], { type: "application/json" });
-              const url = URL.createObjectURL(blob);
-              const a = document.createElement("a");
-              a.href = url;
-              a.download = `${slug || "theme"}.pgtheme.json`;
-              document.body.appendChild(a);
-              a.click();
-              document.body.removeChild(a);
-              URL.revokeObjectURL(url);
-            }}
-          >
-            Export draft
-          </PGButton>
-          <div style={{ flex: 1 }} />
-          <PGButton size="sm" variant="ghost" onClick={handleCancel}>
-            Cancel
-          </PGButton>
-          <PGButton size="sm" variant="primary" icon="check" onClick={handleSave}>
-            {mode === "new" ? "Create theme" : "Save changes"}
-          </PGButton>
-        </footer>
-      </div>
-    </div>
-  );
-}
-
-function ColorEditor({
-  colors,
-  onPatch,
-}: {
-  colors: ThemeColors;
-  onPatch: (p: Partial<ThemeColors>) => void;
-}) {
-  const groups: Array<{
-    title: string;
-    group: "background" | "foreground" | "border" | "accent" | "logo";
-  }> = [
-    { title: "Backgrounds", group: "background" },
-    { title: "Text", group: "foreground" },
-    { title: "Borders", group: "border" },
-    { title: "Accent", group: "accent" },
-    { title: "Logo", group: "logo" },
-  ];
-
-  return (
-    <div style={{ padding: "14px 16px 18px" }}>
-      {groups.map((g) => (
-        <div key={g.group} style={{ marginTop: g.group === "background" ? 0 : 16 }}>
-          <div
-            style={{
-              fontFamily: "var(--font-mono)",
-              fontSize: "var(--fs-10)",
-              textTransform: "uppercase",
-              letterSpacing: "0.05em",
-              color: "var(--fg-2)",
-              fontWeight: 600,
-              marginBottom: 6,
-            }}
-          >
-            {g.title}
-          </div>
-          <div
-            style={{
-              display: "grid",
-              gridTemplateColumns: "repeat(auto-fill, minmax(240px, 1fr))",
-              gap: 8,
-            }}
-          >
-            {THEME_COLOR_FIELDS.filter((f) => f.group === g.group).map((f) => (
-              <ColorField
-                key={f.key}
-                label={f.label}
-                hint={f.hint}
-                value={colors[f.key]}
-                onChange={(v) =>
-                  onPatch({ [f.key]: v } as Partial<ThemeColors>)
-                }
-              />
-            ))}
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function ColorField({
-  label,
-  hint,
-  value,
-  onChange,
-}: {
-  label: string;
-  hint?: string;
-  value: string;
-  onChange: (v: string) => void;
-}) {
-  const [draft, setDraft] = React.useState(value);
-  React.useEffect(() => setDraft(value), [value]);
-
-  const commitHex = (v: string) => {
-    const normalized = normalizeHex(v);
-    if (!normalized) return;
-    onChange(normalized);
-  };
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        alignItems: "center",
-        gap: 8,
-        padding: "6px 8px",
-        border: "1px solid var(--border-0)",
-        borderRadius: "var(--r-3)",
-        background: "var(--bg-1)",
-      }}
-      title={hint}
-    >
-      <label
-        style={{
-          position: "relative",
-          width: 28,
-          height: 28,
-          borderRadius: "var(--r-3)",
-          border: "1px solid var(--border-1)",
-          background: value,
-          cursor: "pointer",
-          flexShrink: 0,
-          overflow: "hidden",
-          boxShadow: "inset 0 0 0 1px rgba(0,0,0,0.15)",
-        }}
-      >
-        <input
-          type="color"
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          style={{
-            position: "absolute",
-            inset: 0,
-            opacity: 0,
-            cursor: "pointer",
-            border: "none",
-          }}
-        />
-      </label>
-      <div style={{ flex: 1, minWidth: 0 }}>
-        <div
-          style={{
-            fontSize: "var(--fs-11)",
-            color: "var(--fg-1)",
-            fontWeight: 500,
-            overflow: "hidden",
-            textOverflow: "ellipsis",
-            whiteSpace: "nowrap",
-          }}
-        >
-          {label}
-        </div>
-        <div style={{ marginTop: 2 }}>
-          <input
-            value={draft}
-            onChange={(e) => setDraft(e.target.value)}
-            onBlur={() => commitHex(draft)}
-            onKeyDown={(e) => {
-              if (e.key === "Enter") {
-                commitHex(draft);
-                (e.target as HTMLInputElement).blur();
-              }
-              if (e.key === "Escape") {
-                setDraft(value);
-                (e.target as HTMLInputElement).blur();
-              }
-            }}
-            spellCheck={false}
-            style={{
-              width: "100%",
-              background: "transparent",
-              border: "none",
-              outline: "none",
-              color: "var(--fg-2)",
-              fontFamily: "var(--font-mono)",
-              fontSize: "var(--fs-11)",
-              padding: 0,
-            }}
-          />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function normalizeHex(v: string): string | null {
-  const raw = v.trim().toLowerCase().replace(/^#/, "");
-  if (/^[0-9a-f]{3}$/.test(raw)) {
-    return `#${raw
-      .split("")
-      .map((ch) => ch + ch)
-      .join("")}`;
-  }
-  if (/^[0-9a-f]{6}$/.test(raw)) return `#${raw}`;
-  return null;
-}
-
